geo-platform/agents/src/requirements.txt
# GEO Agents Requirements
pydantic>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
python-multipart>=0.0.6

# LLM Integration
anthropic>=0.34.0
python-dotenv>=1.0.0
